backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[argon2]
bcrypt==4.0.1
python-jose
pymysql
